# G7 Rapid Reconnect - Faster consistent user friendly behavior