--- a/app/src/main/java/com/eveningoutpost/dexdrip/g5model/Ob1G5StateMachine.java
+++ b/app/src/main/java/com/eveningoutpost/dexdrip/g5model/Ob1G5StateMachine.java
@@ -888,13 +888,14 @@ public class Ob1G5StateMachine {
                             } else {
                                 parent.msg("Invalid Glucose");
                             }
-                            UserError.Log.d(TAG, "Rapid Reconnect is active: " + Ob1G5CollectionService.rapid_reconnect);
-                            if (Ob1G5CollectionService.rapid_reconnect) { // Manage wake frequency after pairing for G7
-                                Ob1G5CollectionService.rapid_reconnect_count++;
-                                UserError.Log.e(TAG, "# of Rapid Reconnect handshakes: " + Ob1G5CollectionService.rapid_reconnect_count);
-                                if (Ob1G5CollectionService.rapid_reconnect_count > 2) {
+                            if (Ob1G5CollectionService.rapid_reconnect_transition) { // Manage wake frequency after pairing with G7
+                                Ob1G5CollectionService.rapid_reconnect_transition_count++; // We have had a handshake.  Therefore, we increment the count.
+                                UserError.Log.e(TAG, "# of Rapid Reconnect handshakes: " + Ob1G5CollectionService.rapid_reconnect_transition_count);
+                                if (Ob1G5CollectionService.rapid_reconnect_transition_count == 2) {
                                     UserError.Log.e(TAG, "Back to waking once every 5 minutes ");
-                                    Ob1G5CollectionService.rapid_reconnect = false;
+                                }
+                                if (Ob1G5CollectionService.rapid_reconnect_transition_count > 2) {
+                                    Ob1G5CollectionService.rapid_reconnect_transition = false;
                                 }
                             }
                             break;