This PR makes the user experience with G7 and xDrip consistent regardless of the random nature caused by the G7 Rapid Reconnect function:
  
| Problem | After this PR |  
| -------- | --------------- |  
| There is an 80% chance xDrip will miss the subsequent transmission to a pairing | xDrip is guaranteed to capture the subsequent transmission to pairing  |  
| There is 80% chance that the next capture after pairing occurs in 6-9 minutes | There is 80% chance that the next capture after pairing occurs in 1-4 minutes |  
| An error message may be shown after pairing | No error message will be shown after pairing |  
| xDrip may miss the second transmission after pairing | xDrip is guaranteed to capture the second transmission after pairing |  
| There is 20% chance that xDrip will pair on the once every 5 minute time grid.  Even then, xDrip may miss the second handshake after the pairing handshake. | There is 20% chance xDrip will pair on the once every 5 minute time grid.  xDrip is guaranteed to capture the second handshake after the pairing handshake. |  
  
---  
  
**What activates Rapid Reconnect Transition mode?**
It is automatically activated when xDrip pairs with a G7.  
I refer to this in the logs as the first Rapid Reconnect Transition mode handshake.
  
---  
  
**What happens in Rapid Reconnect Transition mode?**  
xDrip wakes once a minute.  
The Hunting Transmitter note is shown on the Dex status page after pair  approval until a subsequent handshake regardless of any event.
Subsequent to the pair request approval, after xDrip completes another handshake, scan always mode is activated.  
No brain status error is shown on the Dex status page.
  
---  
    
**What disables Rapid Reconnect Transition mode?**  
When either one of the following two conditions is satisfied, the Rapid Reconnect Transition mode is disabled.  
  
1- Two handshakes subsequent to the pair request approval handshake
2- 10 minutes after pair request approval  
  
---  
  
**Could it fail?**  
If the user walks away from the phone after pairing before two subsequent handshakes, the intended process will not complete.  
But, the user can just wait for xDrip to do its job.  Alternatively, the user can forget the device from the Bluetooth list and temporarily change the transmitter ID and wait for a pair request for the Rapid Reconnect Transition mode to activate again.  
  
---  
  
**Test**  
  
The following tests were performed on an Android 11 Motorola Stylus phone.
Before every test, xDrip was uninstalled and all devices starting with DXCM were deleted from the Bluetooth list. 
Then, the version under test was installed.  Then, xDrip battery optimization was disabled.  After that, source wizard was used to choose G7, background location access was granted, and the pairing code was entered.  

For both tests, the transmitter was disconnected more than 15 minutes prior and it was confirmed that it was transmitting once a minute.
For both tests, it was arranged for pairing to occur one 1-minute cycle before the 5-minute cycle transmission.  
For both tests, collection was enabled about 30 seconds before the transmission cycle start.  

Screens recordings in the following table are sped up by a factor of 16.  
| Before | After |  
| ------- | ------ |  
| https://github.com/user-attachments/assets/0e8c0306-fb9f-44ff-bb81-918fc7b0651d | https://github.com/user-attachments/assets/02a373ef-8c51-4f3c-a786-ecc06fd4c608 |  
| ![Before](https://github.com/user-attachments/assets/31e5455d-06c7-4114-9d53-f0bba37b98f4) | ![After](https://github.com/user-attachments/assets/8167ece4-1d22-4582-91ed-9e6665c3f409) |  
  
---  
  
If the user waits patiently for several read cycles for xDrip to enable always scan mode, it will eventually discover the correct time grid.
The problem is that many users will intervene thinking that something is wrong considering the possibility of the error message and the longer periods of time with no connectivity.  So, they restart the phone or restart collector etc.  Sometimes, these will succeed.  

The guide asks the user to read this: https://navid200.github.io/xDrip/docs/Dexcom/G7OnceAMinuteScan.html
If user follows those instructions, the outcome will be identical to what this PR does.
The problem is that you cannot expect everyone to read all the pages of the guide. 
This PR does the right thing with no need for any human intervention other than keeping the old sensors away and staying with the phone for up to 10 minutes after pairing. 

This PR ensures connectivity as fast as possible taking into account everything that could possibly go wrong.