@Navid200 2024-12-26
Let's see what happens if we use our latest Nightly and pair on the wrong grid without human intervention:
  
![Screenshot_20241225-225334](https://github.com/user-attachments/assets/bef089d8-7fe6-4a65-8910-fe8c07c4293f)  ![Screenshot_20241225-225309](https://github.com/user-attachments/assets/8738f154-c67f-41de-8cc8-d48642a141f5)
  
After 1 hour since pairing, there is still no proper connectivity.  As G7 goes in and out of Rapid Reconnect, xDrip connects.  But, it still has not found the correct time grid.
  
You can see what I mean by not having connected properly by looking at the time values on the left in the logs.  The time gap is never 5 minutes.  
  
Of course, this can be remedied if these instructions are followed:
https://github.com/NightscoutFoundation/xDrip/discussions/3813  
But, that requires human intervention.  

Now, lets see what happens if we do exactly the same thing after this PR.