@Navid200 2025-01-16
@jamorham I'm sorry about the long description.  
You don't have to read the whole thing.  Please just tell me what you think or if you have any question.