@Navid200 2025-01-19
On cannot recreate the problem with my Pixel.