@jamorham 2025-01-24
Thanks for the very detailed PR. I will have to study it in some detail because making changes to the timing aspect is a very significant change which has to be handled really well. I appreciate there are some inconsistencies already and it is definitely worth a closer look to see if things can be improved but equally I don't find consistent behavior from the transmitter side which makes testing harder.