@Navid200 2025-01-31
Before and after for a fully functional G7:
| Before | After |  
| ------- | ------ |  
| https://github.com/user-attachments/assets/d10b761d-509c-41a5-9a45-f60761fef97a | https://github.com/user-attachments/assets/f1c94e54-5918-4162-91ab-3bd364ffeba0 |  
| ![Screenshot_20250131-104557](https://github.com/user-attachments/assets/a3eed6c6-0043-42e0-a82a-dbe4254e80a1) | ![Screenshot_20250131-112400](https://github.com/user-attachments/assets/11b1a4fa-932e-4147-b94f-8a628b74386c) |