@Navid200 2025-02-01
> Thanks for the very detailed PR. I will have to study it in some detail because making changes to the timing aspect is a very significant change which has to be handled really well. I appreciate there are some inconsistencies already and it is definitely worth a closer look to see if things can be improved but equally I don't find consistent behavior from the transmitter side which makes testing harder.

Thanks
This is a critical change and I appreciate a critical review.

I hope I have not misunderstood the comment about testing challenge.
I identify the timing of the transmitter and write it down.  It could be something like: 00:02:50
This implies 00:07:50 as well as the time grid is cycles at 5 minutes.

Now that I have that, I know the rapid reconnect timings will be 00:3:50, 00:04:50, 00:05:50 and 00:06:50.  Of course, there are another 4 at 00:08:50 ...
Having the rapid reconnect timings, I connect to the transmitter.  Then, I disconnect from it and wait 15 minutes.  This guarantees that it is in rapid reconnect mode.  I can also verify using a scanner.

I install xDrip on the test phone and keep collection disabled during all scans.  Each scan lasts about 17 seconds.
When I know there is no scan, I enable G7, and scan the QR code etc.  
When xDrip is ready, I time enabling collection such that xDrip is scanning 15 seconds before the start of the particular rapid reconnect slot I want to test.
I use the exact same slot for my before and after tests for a valid comparison.