@Navid200 2025-04-25
This can remove error messages that should not be shown because they are not real errors and are simply caused by the rapid reconnect behavior of G7.But, relatively speaking, I would look at a few other PRs before this.For example, there is a PR that upload treatments to multiple Nightscout sites.I followed your advice and believe that addressed your concerns.There is another PR that allows a delay to be added to the dex share follower.I hope you can consider those as they bought have been waiting for s while for a review.On Apr 25, 2025 5:45 a.m., JamOrHam ***@***.***> wrote:jamorham left a comment (NightscoutFoundation/xDrip#3829)
Do you think this is still needed? I thought that other changes had improved this overall. Please let me know what you think.

—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you authored the thread.Message ID: ***@***.***>